Det skal ikke være mulig å innvilge en behandling hvis ikke alle vilkår er oppfylt.
