+++ src/frontend/Komponenter/Behandling/VedtakOgBeregning/VedtakOgBeregning.tsx
@@ -7,19 +7,24 @@ import VedtaksresultatSwitch from './VedtaksresultatSwitch';
 import SelectVedtaksresultat from './SelectVedtaksresultat';
 import { Behandling } from '../../../App/typer/fagsak';
 import { useHentVedtak } from '../../../App/hooks/useHentVedtak';
+import { IVilkår, IVurdering, Vilkårsresultat } from '../Inngangsvilkår/vilkår';
 
 interface Props {
     behandling: Behandling;
+    vilkår: IVilkår;
 }
 
 const Wrapper = styled.div`
     padding: 1rem 2rem;
 `;
-const VedtakOgBeregning: FC<Props> = ({ behandling }) => {
+const VedtakOgBeregning: FC<Props> = ({ behandling, vilkår }) => {
     const behandlingId = behandling.id;
     const [resultatType, settResultatType] = useState<EBehandlingResultat>();
     const { vedtak, hentVedtak } = useHentVedtak(behandlingId);
 
+    const alleOppfylt = vilkår.vurderinger.every(
+        (vurdering: IVurdering) => vurdering.resultat === Vilkårsresultat.OPPFYLT
+    );
     useEffect(() => {
         hentVedtak();
     }, [hentVedtak]);
@@ -39,6 +44,7 @@ const VedtakOgBeregning: FC<Props> = ({ behandling }) => {
                             behandling={behandling}
                             resultatType={resultatType}
                             settResultatType={settResultatType}
+                            alleOppfylt={alleOppfylt}
                         />
                         <VedtaksresultatSwitch
                             vedtaksresultatType={resultatType}
