+++ src/frontend/Komponenter/Behandling/VedtakOgBeregning/VedtakOgBeregningSide.tsx
@@ -1,4 +1,4 @@
-import React, { FC } from 'react';
+import React, { FC, useCallback, useEffect } from 'react';
 import { useBehandling } from '../../../App/context/BehandlingContext';
 import { Steg } from '../Høyremeny/Steg';
 import { Søknadsdatoer } from './Søknadsdatoer';
@@ -8,6 +8,7 @@ import AlertStripe from 'nav-frontend-alertstriper';
 import { VilkårsresultatOppsummering } from '../Vilkårresultat/VilkårsresultatOppsummering';
 import VedtakOgBeregning from './VedtakOgBeregning';
 import DataViewer from '../../../Felles/DataViewer/DataViewer';
+import { useHentVilkår } from '../../../App/hooks/useHentVilkår';
 import { Behandlingsårsak } from '../../../App/typer/Behandlingsårsak';
 
 const AlertStripeLeft = styled(AlertStripe)`
@@ -24,19 +25,29 @@ const AlertStripeIkkeFerdigBehandletVilkår = (): JSX.Element => (
 export const VedtakOgBeregningSide: FC<{ behandlingId: string }> = ({ behandlingId }) => {
     const { behandling } = useBehandling();
 
+    const { vilkår, hentVilkår } = useHentVilkår();
+
+    const hentVilkårCallback = useCallback(() => {
+        hentVilkår(behandlingId);
+    }, [behandlingId, hentVilkår]);
+
+    useEffect(() => {
+        hentVilkårCallback();
+    }, [hentVilkårCallback]);
+
     return (
-        <DataViewer response={{ behandling }}>
-            {({ behandling }) => {
+        <DataViewer response={{ behandling, vilkår }}>
+            {({ behandling, vilkår }) => {
                 const skalViseSøknadsdata = behandling.behandlingsårsak === Behandlingsårsak.SØKNAD;
 
                 return (
                     <>
-                        <VilkårsresultatOppsummering behandlingId={behandlingId} />
+                        <VilkårsresultatOppsummering vilkår={vilkår} />
                         {skalViseSøknadsdata && <Søknadsdatoer behandlingId={behandlingId} />}
                         {behandling.steg === Steg.VILKÅR ? (
                             <AlertStripeIkkeFerdigBehandletVilkår />
                         ) : (
-                            <VedtakOgBeregning behandling={behandling} />
+                            <VedtakOgBeregning behandling={behandling} vilkår={vilkår} />
                         )}
                     </>
                 );
